# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nothing to see here",
        "title": "An API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auths": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the auths",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auths"
                ],
                "summary": "Get all auths",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/backup": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Backup the cowrie database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Backup the cowrie database",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/report": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate a report",
                "tags": [
                    "report"
                ],
                "summary": "Generate a report",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/restore/{filename}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Restore the cowrie database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restore"
                ],
                "summary": "Restore the cowrie database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sessions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the sessions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get all sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit the number of sessions returned",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include failed logins",
                        "name": "include_failed_logins",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Please provide a valid api token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}